having trouble with parts of the deserialization of the JSON
