+++ app/build.gradle.kts
@@ -269,6 +269,8 @@ dependencies {
     androidTestImplementation("androidx.arch.core:core-testing:2.1.0")
 
     testImplementation("androidx.arch.core:core-testing:2.2.0")
+    
+    implementation("com.squareup.okhttp3:logging-interceptor:3.12.0")
 }
 
 tasks.register("jacocoTestReport", JacocoReport::class) {
