+++ app/src/main/java/com/github/wanderwise_inc/app/ui/creation/steps/CreateItineraryMap.kt
@@ -194,7 +194,7 @@ fun LocationSelector(
                 onDone = { onSearch(location1) },
                 onSearch = { onSearch(location1) }
             ),
-            
+            singleLine = true
         )
       }
