+++ app/src/test/java/com/github/wanderwise_inc/app/network/LocationsApiServiceKtTest.kt
@@ -1,0 +1,7 @@
+package com.github.wanderwise_inc.app.network
+
+import org.junit.jupiter.api.Assertions.*
+
+import junit.framework.TestCase
+
+class LocationsApiServiceKtTest : TestCase()
